requirements.txt
tensorflow==2.13.0
numpy==1.23.5
opencv-python==4.8.1.78
Pillow==9.3.0
streamlit==1.26.0
requires.io

